freelance-portfolio/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
fastapi-mail
gunicorn
